Moves permission arrays into shared permissions file.
